Enabled quick testing
