+++ MiddleTier/cb.json
@@ -1,0 +1,1 @@
+{"numberOfFaces":1,"faces":[{"emotions":[{"emotion":"sadness","value":0.9966155},{"emotion":"disgust","value":0.00333234319},{"emotion":"anger","value":0.0000337475358}]}],"colorEntries":[{"Percent":53.9,"Color":"neutral"},{"Percent":27.33,"Color":"neutral"}]}
