+++ MiddleTier/package.json
@@ -17,6 +17,7 @@
     "express": "^4.13.4",
     "express-session": "^1.13.0",
     "imagecolors": "^0.10.5",
+    "jsonfile": "^2.2.3",
     "mongoose": "^4.4.0",
     "morgan": "^1.6.1",
     "multer": "^1.1.0",
