+++ MiddleTier/solo.json
@@ -1,0 +1,1 @@
+{"numberOfFaces":3,"faces":[{"emotions":[{"emotion":"neutral","value":0.9637481},{"emotion":"happiness","value":0.03260274},{"emotion":"contempt","value":0.00318247732}]},{"emotions":[{"emotion":"happiness","value":1},{"emotion":"neutral","value":2.14160156e-9},{"emotion":"disgust","value":1.31224476e-9}]},{"emotions":[{"emotion":"happiness","value":0.9779488},{"emotion":"neutral","value":0.01760259},{"emotion":"contempt","value":0.003959531}]}],"colorEntries":[{"Percent":28.45,"Color":"purple"},{"Percent":26.28,"Color":"blue"}]}
